responsive homepage

+++ routes/user_deckbuilder_routes.py
@@ -13,18 +13,6 @@ from __main__ import turbo
 def buy_list_page(username):
     create_list_form = CreateList()
     if username == current_user.username :
-        #create_list_form = CreateList()
-        #if create_list_form.validate_on_submit():
-        #    print('on rentre')
-        #    name = cleanhtml(create_list_form.new_list.data)
-        #    id_type=cleanhtml(create_list_form.id_deckbuilde.data)
-        #    date = datetime.datetime.now()
-        #    id_user = current_user.id      
-        #    create = create_list([name, date, id_type, id_user])
-        #    if create:
-        #        flash(f"Votre liste a bien été ajoutée.", "succès")
-        #    else:
-        #        flash(f"Création de cette liste impossible.", "erreur")
         buy_lists = db.session.query(User, DeckBuilder, DeckBuilder_Card, Card).join(DeckBuilder, User.id == DeckBuilder.id_user).join(
             DeckBuilder_Card, DeckBuilder.id == DeckBuilder_Card.id_deckbuilder).join(
                 Card, Card.id == DeckBuilder_Card.id_card).filter(DeckBuilder.id_user == current_user.id).filter(
@@ -69,7 +57,7 @@ def buy_list_page(username):
 @is_logged
 def create_deckbuilder_list(id_type):
     create_list_form = CreateList()
-    if create_list_form.validate_on_submit():
+    if request=="POST" and create_list_form.validate_on_submit():
         name = cleanhtml(create_list_form.new_list.data)
         date = datetime.datetime.now()
         id_user = current_user.id      
@@ -90,8 +78,6 @@ def create_deckbuilder_list(id_type):
     if id_type == 3 :
         return redirect(url_for('deckbuilder_list_page', username=current_user.username))
 
-
-
 @app.route('/user/add_to_buy_list', methods = ['POST', 'GET'])
 @is_logged
 def add_to_buy():
@@ -105,7 +91,6 @@ def add_to_buy():
                 flash(f"{card_name.capitalize()} a bien été rajoutée.", "succès")
             else:
                 flash(f"Problème lors de l'enregistrement de {card_name.capitalize()}.", "erreur")
-
         else :
             flash(f"Problème lors de l'enregistrement.", "erreur")
     return redirect(url_for('buy_list_page', username=current_user.username))
@@ -113,37 +98,40 @@ def add_to_buy():
 @app.route('/user/delete_page_list', methods=['POST'])
 @is_logged
 def delete_list():   
-    id = cleanhtml(request.form['delete-id'])
-    list_cards_to_delete = DeckBuilder_Card.query.filter_by(id_deckbuilder=id).all()
-    if list_cards_to_delete != []:
-        for card_to_delete in list_cards_to_delete:
-            db.session.delete(card_to_delete)
-        db.session.commit()
-    list_to_delete = DeckBuilder.query.filter_by(id=id).first()
-    if list_to_delete:
-        id_type = list_to_delete.id_type
-        list_name = list_to_delete.name
-        db.session.delete(list_to_delete)
-        db.session.commit()
-        flash(f"{list_name} a bien été supprimée.", "succès")
-    else:
-        flash(f"Suppression de la liste impossible.", "erreur")
-    if id_type == 1 :
-        return redirect(url_for('buy_list_page', username=current_user.username))
-    if id_type == 2 :
-        return redirect(url_for('sell_list_page', username=current_user.username))
-    if id_type == 3 :
-        return redirect(url_for('deckbuilder_list_page', username=current_user.username))
+    if request.method == 'POST':
+        id = cleanhtml(request.form['delete-id'])
+        list_cards_to_delete = DeckBuilder_Card.query.filter_by(id_deckbuilder=id).all()
+        if list_cards_to_delete != []:
+            for card_to_delete in list_cards_to_delete:
+                db.session.delete(card_to_delete)
+            db.session.commit()
+        list_to_delete = DeckBuilder.query.filter_by(id=id).first()
+        if list_to_delete:
+            id_type = list_to_delete.id_type
+            list_name = list_to_delete.name
+            db.session.delete(list_to_delete)
+            db.session.commit()
+            flash(f"{list_name} a bien été supprimée.", "succès")
+        else:
+            flash(f"Suppression de la liste impossible.", "erreur")
+        if id_type == 1 :
+            return redirect(url_for('buy_list_page', username=current_user.username))
+        if id_type == 2 :
+            return redirect(url_for('sell_list_page', username=current_user.username))
+        if id_type == 3 :
+            return redirect(url_for('deckbuilder_list_page', username=current_user.username))
+    return redirect(url_for('home'))
 
 
 @app.route('/user/delete_buy_card_list/<int:card_id>/<int:deckbuilder_id>', methods=['POST'])
 @is_logged
 def delete_buy_card_list(card_id, deckbuilder_id):
-    card_to_delete = DeckBuilder_Card.query.filter(DeckBuilder_Card.id_card == card_id).filter(DeckBuilder_Card.id_deckbuilder == deckbuilder_id).first()
-    if card_to_delete:
-        db.session.delete(card_to_delete)
-        db.session.commit()
-        return redirect(url_for('buy_list_page', username=current_user.username))
+    if request.method == 'POST':
+        card_to_delete = DeckBuilder_Card.query.filter(DeckBuilder_Card.id_card == card_id).filter(DeckBuilder_Card.id_deckbuilder == deckbuilder_id).first()
+        if card_to_delete:
+            db.session.delete(card_to_delete)
+            db.session.commit()
+            return redirect(url_for('buy_list_page', username=current_user.username))
     return redirect(url_for('home', username=current_user.username))
 
 
@@ -195,22 +183,25 @@ def add_to_sell():
         sells = add_card_to_sell_list([id_deckbuilder, card_name, nb, edition, creation, language, rarity, quality, price])
         if sells:
             flash(f"{card_name.capitalize()} a bien été rajoutée.", "succès")
-        else:
+        else :
             flash(f"Problème lors de l'enregistrement de {card_name.capitalize()}.", "erreur")
-        return redirect(url_for('sell_list_page', username=current_user.username))
+    if request.method == 'POST' and not sell_form.validate_on_submit():
+        flash("Enregistrement impossible. Merci de bien compléter tout les champs.", "erreur")
+    return redirect(url_for('sell_list_page', username=current_user.username))
   
 @app.route('/user/delete_sell_card_list', methods=['POST'])
 @is_logged
 def delete_sell_card_list():
-    card_id = cleanhtml(request.form['card-id'])
-    deckbuilder_id = cleanhtml(request.form['deck-id'])
-    card_to_delete = DeckBuilder_Card.query.filter(DeckBuilder_Card.id_card == card_id).filter(DeckBuilder_Card.id_deckbuilder == deckbuilder_id).first()
-    card_name = card_to_delete.name
-    if card_to_delete:
-        db.session.delete(card_to_delete)
-        db.session.commit()
-        flash(f"{card_to_delete.capitalize()} a bien été supprimée.","succès")
-    flash(f"{card_to_delete.capitalize()} n'a pas été supprimée.", "erreur")
+    if request.method == 'POST':
+        card_id = cleanhtml(request.form['card-id'])
+        deckbuilder_id = cleanhtml(request.form['deck-id'])
+        card_to_delete = DeckBuilder_Card.query.filter(DeckBuilder_Card.id_card == card_id).filter(DeckBuilder_Card.id_deckbuilder == deckbuilder_id).first()
+        card_name = card_to_delete.name
+        if card_to_delete:
+            db.session.delete(card_to_delete)
+            db.session.commit()
+            flash(f"{card_to_delete.capitalize()} a bien été supprimée.","succès")
+        flash(f"{card_to_delete.capitalize()} n'a pas été supprimée.", "erreur")
     return redirect(url_for('sell_list_page', username=current_user.username))
